Revert "make default endpoint not hidden"
This reverts commit 5deeb030833452d0babb2179d98a6d323ff15350.
